+++ resources/js/Pages/Dashboard/Show.svelte
@@ -10,6 +10,7 @@
     /* svelte-ignore unused-export-let */
     export let flash = {};
     export let task;
+    export let client_d;
     let editTask = false;
     let editReplyTask = false;
     let replyNow = false;
@@ -142,10 +143,12 @@
             <div class="text-green-500 font-bold">{flash.message}</div>
         {/if}
 
-        <div class="flex justify-end">
-            <a use:inertia href="/dashboard/task/{task.slug}/edit" class="inline-block mt-2 text-blue-500 border border-gray-200 px-4 py-1 rounded-lg bg-white">Edit Task</a>
-            <button type="button" use:inertia="{{ href: `/dashboard/task/${task.id}`, method: 'delete', replace: true, }}" class="inline-block mt-2 text-white border border-gray-200 px-4 py-1 rounded-lg bg-red-400 ml-4">Delete</button>
-        </div>
+        {#if task.user_id === client_d}
+            <div class="flex justify-end">
+                <a use:inertia href="/dashboard/task/{task.slug}/edit" class="inline-block mt-2 text-blue-500 border border-gray-200 px-4 py-1 rounded-lg bg-white">Edit Task</a>
+                <button type="button" use:inertia="{{ href: `/dashboard/task/${task.id}`, method: 'delete', replace: true, }}" class="inline-block mt-2 text-white border border-gray-200 px-4 py-1 rounded-lg bg-red-400 ml-4">Delete</button>
+            </div>
+        {/if}
         <div class="space-y-4 mt-5">
             <div class="relative border border-gray-200 rounded p-4 my-2">
                 <div class="flex justify-between items-center">
